refactor(render): satisfy viewport quality gates HORO-293 #293 [RND-001.4]
