+++ src/editor/renderer/opengl/OpenGLViewportResourceBridge.h
@@ -7,17 +7,17 @@ namespace Horo::Editor {
     class OpenGLViewportResourceBridge final {
     public:
         /** @brief Binds a ready generic mesh's backend-private vertex input object. */
-        [[nodiscard]] Result<void> BindMesh(const Render::RenderFrontend &frontend, Render::RenderMeshHandle mesh) const;
+        [[nodiscard]] static Result<void> BindMesh(const Render::RenderFrontend &frontend, Render::RenderMeshHandle mesh);
 
         /** @brief Binds a ready generic offscreen render target. */
-        [[nodiscard]] Result<void> BindRenderTarget(const Render::RenderFrontend &frontend, Render::RenderTargetHandle target) const;
+        [[nodiscard]] static Result<void> BindRenderTarget(const Render::RenderFrontend &frontend, Render::RenderTargetHandle target);
 
         /** @brief Binds a ready generic texture view to an OpenGL texture unit slot. */
-        [[nodiscard]] Result<void> BindTexture(const Render::RenderFrontend &frontend, Render::RenderTextureViewHandle view,
-                                               std::uint32_t unit) const;
+        [[nodiscard]] static Result<void> BindTexture(const Render::RenderFrontend &frontend, Render::RenderTextureViewHandle view,
+                                                      std::uint32_t unit);
 
         /** @brief Returns the opaque GUI identity derived from a ready generic texture view. */
-        [[nodiscard]] Result<std::uintptr_t> EditorImageIdentity(const Render::RenderFrontend &frontend,
-                                                                 Render::RenderTextureViewHandle view) const;
+        [[nodiscard]] static Result<std::uintptr_t> EditorImageIdentity(const Render::RenderFrontend &frontend,
+                                                                        Render::RenderTextureViewHandle view);
     };
 }  // namespace Horo::Editor
